+++ mitmproxy/tools/console/consoleaddons.py
@@ -118,7 +118,7 @@ class ConsoleAddon:
             "Strip trailing newlines from edited request/response bodies.",
         )
         loader.add_option(
-            "console_quickhelp",
+            "console_quickhelp_visible",
             bool,
             True,
             "Show the quick help bar at the bottom of the console.",